Data is getting added and deleted simultaneously in firebase for profile

+++ src/app/profile/profile.page.ts
@@ -42,7 +42,7 @@ export class ProfilePage implements OnInit {
   UIDfromFirebase:any;
   u:any;
   t:any;
-
+  DB:any;
   constructor(
     private authservice: AuthService,
     private afAuth: AngularFireAuth,
@@ -60,17 +60,55 @@ export class ProfilePage implements OnInit {
     
              };
 
-  view()     {
-                    this.delete();
+
+             check (){
+                 this.view();
              };
+
+
+
+
+  view()     {
+    
+                var database1 = firebase.database();
+                var refxx     = database1.ref('users/');
+                refxx.on('value',((gotdata1)=>{
+                  var db2 = gotdata1.val();
+                  console.log(db2);
+                  if(db2 !==null){
+                                    var keyss= Object.keys(db2);
+                                    console.log(keyss);
+                                         {
+                                                  if(keyss.length>0)
+                                                   this.delete();
+                    
+                 
+                                          }
+                                  }
+                 else{
+                         refxx.off('value');
+                         console.log("Inside updating data");
+                         this.update();
+                
+                      }
+                 
+                                                }
+                  )
+                  ) ;
+                
+                };
+
+
+
 
   delete()   {
                     var database = firebase.database();
-                    var refx = database.ref("users/");
-                    refx.on("value", (gotdata) => {
+                    var refx = database.ref('users/');         
+                    refx.on('value', ((gotdata) => {
                                                       this.DBdata = gotdata.val();
+                                                      console.log("val ran");                                              
                                                       var key = Object.keys(this.DBdata);
-                                                      console.log(key);
+                                                      console.log(key);                                                 
                                                             for (var i = 0; i < key.length; i++)
                                                              {
                                                                 this.k = key[i];
@@ -78,21 +116,30 @@ export class ProfilePage implements OnInit {
                                                                 console.log(this.userId);
                                                                 console.log(this.UIDfromFirebase);
                                                                 if (this.userId == this.UIDfromFirebase) {
-                                                                    var z = key[i];
+                                                                 var counter = 1;
+                                                                 var m = i ;
+                                                                } else{
+                                                                  refx.off('value');
+                                                                  this.update();
+                                                                }
+                                                              }
+                                                                {
+                                                                  if (counter ==1){
+                                                                    var z = key[m];
+                                                                    console.log(z);
                                                                     var removeolddetails = firebase.database().ref(`users/${z}`);
                                                                     removeolddetails.remove().then(function () {
                                                                           console.log("Data Removed");
                                                                                                   }).then(function(){
+                                                                                                    refx.off('value');
                                                                                                     this.update();
                                                                                                   });
                                                                 } 
-
-
-                                                             }
-     
-                                                  }
+                                                              }                                                 
+                                                  })
                              );
 
+                             refx.off('value');
              };
 
     
@@ -100,38 +147,36 @@ export class ProfilePage implements OnInit {
   
 
    update() {
-                     const x = this.username;
+                     
                      console.log("username - ", this.username);
                      console.log("Mobile - ", this.mobile);
-                     this.db.list(`users/`).push({
+                     this.db.list('users/').push({
                      uid: this.userId,
                      email: this.email,
                      name: this.username,
                      mobile: this.mobile,
-                                                 });
-                               
+                                                 });                            
                      var database = firebase.database();
-                     var refx=database.ref("users/");
-                     refx.on("value",(fetchdata)=>{
-                                                    var DB = fetchdata.val();
-                                                    var key = Object.keys(DB);
+                     var refx=database.ref('users/');
+                     refx.on('value',((fetchdata)=>{
+                                                    this.DB = fetchdata.val();
+                                                    var key = Object.keys(this.DB);
+                                                    
                                                       for (var i =0;i<key.length;i++){
                                                               this.t =key[i];
-                                                             this.u = this.DBdata[this.t].uid;
+                                                              
+                                                             this.u = this.DB[this.t].uid;
+                                                             console.log(this.u);
                                                              console.log(this.userId,this.u);
                                                                 if(this.userId == this.u){
-                                                                      var NameFromFirebase = this.DBdata[this.t].name;
-                                                                      var Mob = this.DBdata[this.t].mobile;
+                                                                      var NameFromFirebase = this.DB[this.t].name;
+                                                                      var Mob = this.DB[this.t].mobile;
                                                                       console.log(NameFromFirebase, Mob);
                                                                                     }
                                                                                       }
-                                                   });
-    
-  
+                                                   }));
 
- 
+                                                   refx.off('value');
             };
-
-
 }
